resources/requirements.txt
pillow==12.0.0
treepoem==3.28.0

# manual/choco install required for treepoem:
# Ghostscript >= 10.0 (https://ghostscript.com/)